feat(state): Define BeaconState (#609)
* feat(client): Calculate GenesisValidatorRoot as we construct the genesis

* fork

* bs

* bet

* bet

* running again

* lint

* gen

* generate

* add comment

* add comment

* gen

* state processor bump

* bet

* types

* use signature

* bet

* rip

* bet

* remove deadcode

* gen chekc

---------

+++ store/blob/blob.go
@@ -78,7 +78,7 @@ func (s *Store) Persist(
 			if err != nil {
 				return err
 			}
-			return s.Set(slot, sc.KzgCommitment, bz)
+			return s.Set(uint64(slot), sc.KzgCommitment, bz)
 		},
 	)...)
 }
